refactor(contexts): Reorganize context modules into structured folders
This commit restructures the application's context management by:
- Moving auth, axios, chatroom, history, and socket contexts to dedicated folders
- Consolidating related Context, Provider, and hook files under `contexts/`
- Creating new GameListContext to centralize game state management
- Updating all import paths throughout the application

This reorganization improves code maintainability by grouping related
functionality together and provides a clearer separation of concerns
for context-based state management.

Related changes:
- Updated 25+ component files with new import paths
- Removed scattered context files from containers/provider/ and hooks/context/
- Enhanced module organization following domain-driven structure
