+++ contexts/auth/AuthProvider.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactNode, useState } from "react";
 import { UserInfo } from "@/requests/users";
-import AuthContext from "@/contexts/AuthContext";
+import AuthContext from "./AuthContext";
 
 type Props = {
   children: ReactNode;
