+++ contexts/auth/useAuth.ts
@@ -1,4 +1,4 @@
-import AuthContext from "@/contexts/AuthContext";
+import AuthContext from "./AuthContext";
 import { useContext } from "react";
 
 const useAuth = () => {
