+++ contexts/axios/useAxios.ts
@@ -1,6 +1,5 @@
+import AxiosContext from "./AxiosContext";
 import { useContext } from "react";
-
-import AxiosContext from "@/contexts/AxiosContext";
 
 const useAxios = () => {
   return useContext(AxiosContext);
