+++ contexts/game/GameListContext.tsx
@@ -1,0 +1,6 @@
+import { type GameType } from "@/requests/games";
+import { createContext } from "react";
+
+const GameListContext = createContext<GameType[]>([]);
+
+export default GameListContext;
