--- features/game/contexts/index.ts
@@ -1,2 +1,0 @@
-export { default as GameListProvider } from "./GameList";
-export * from "./GameList";
