+++ features/game/index.ts
@@ -1,2 +1,1 @@
 export * from "./components";
-export * from "./contexts";
