+++ pages/rooms/index.tsx
@@ -5,7 +5,7 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import Image from "@/components/shared/Image";
 import Modal from "@/components/shared/Modal";
 import Tabs, { TabItemType } from "@/components/shared/Tabs";
-import { useGameList } from "@/features/game";
+import { useGameList } from "@/contexts/game";
 import { JoinLockRoomForm, RoomCard } from "@/features/room";
 import usePagination from "@/hooks/usePagination";
 import useRequest from "@/hooks/useRequest";
