+++ services/socket/types.ts
@@ -7,7 +7,7 @@
  */
 
 import { Socket } from "socket.io-client";
-import { SOCKET_EVENT } from "@/contexts/SocketContext";
+import { SOCKET_EVENT } from "@/contexts/socket";
 
 /**
  * Base interface for all socket services
