+++ services/socket/utils.ts
@@ -1,5 +1,5 @@
 import { Socket } from "socket.io-client";
-import { SOCKET_EVENT } from "@/contexts/SocketContext";
+import { SOCKET_EVENT } from "@/contexts/socket";
 
 /**
  * Type for socket event handler function
